// webapp/src/api/model/grantedDatabaseObjectIdAuthorityDTO.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { IGrantedAuthorityDTO } from './iGrantedAuthorityDTO';
import type { GrantedDatabaseObjectIdAuthorityDTOAllOf } from './grantedDatabaseObjectIdAuthorityDTOAllOf';

export type GrantedDatabaseObjectIdAuthorityDTO = IGrantedAuthorityDTO & GrantedDatabaseObjectIdAuthorityDTOAllOf & Required<Pick<IGrantedAuthorityDTO & GrantedDatabaseObjectIdAuthorityDTOAllOf, 'id' | 'permission'>>;